libs/validator/src/test/java/de/unibi/cebitec/bibiserv/util/validate/impl/RNAMotif_ValidatorTest.java
/*
 * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
 *
 * Copyright 2010 BiBiServ Curator Team, http://bibiserv.cebitec.uni-bielefeld.de,
 * All rights reserved.
 *
 * The contents of this file are subject to the terms of the Common
 * Development and Distribution License("CDDL") (the "License"). You
 * may not use this file except in compliance with the License. You can
 * obtain a copy of the License at http://www.sun.com/cddl/cddl.html
 *
 * See the License for the specific language governing permissions and
 * limitations under the License.  When distributing the software, include
 * this License Header Notice in each file.  If applicable, add the following
 * below the License Header, with the fields enclosed by brackets [] replaced
 *  by your own identifying information:
 *
 * "Portions Copyrighted [year] [name of copyright owner]"
 *
 * Contributor(s):
 *
 */
package de.unibi.cebitec.bibiserv.util.validate.impl;

import de.unibi.cebitec.bibiserv.util.validate.ValidationResult;
import java.io.IOException;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author Sven Hartmeier
 */
public class RNAMotif_ValidatorTest extends AbstractTest{

    String rnaMotiv, rnaMotiverror;

    public RNAMotif_ValidatorTest() {
        try {
            rnaMotiv = readFromResource("/rnamotif_correct.xml");
            rnaMotiverror = readFromResource("/rnamotif_incorrect.xml");
        } catch (IOException e) {
            e.printStackTrace();
            fail(e.toString());
        }
    }

    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @Before
    public void setUp() {
    }

    @After
    public void tearDown() {
    }

    @Test
    public void testRNAMotif() {
        RNAMotif_Validator validator = new RNAMotif_Validator();

        //correct
        ValidationResult vr = validator.validateThis(rnaMotiv);
        assertTrue(vr.getMessage(), vr.isValid());
        //abroken
        vr = validator.validateThis(rnaMotiverror);
        assertTrue(vr.getMessage(), !vr.isValid());

    }

}
